Refactor MoE optimizer modifications
Now part of new upstream MoE code.
